requirements.txt
# WorldScreener Requirements

# Core packages
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Financial data retrieval
yfinance>=0.2.58
requests>=2.25.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
tabulate>=0.8.9
argparse>=1.4.0

# Report generation
jinja2>=3.0.0
xlsxwriter>=3.0.0

# Optional dependencies
# Uncomment to install
# weasyprint>=53.0       # For PDF report generation
# pygal>=2.4.0           # Alternative visualization library
# plotly>=5.3.0          # Interactive visualizations
# dash>=2.0.0            # Web dashboard